Move HTTP module into Insights
