+++ static/app/views/insights/common/components/tableCells/spanIdCell.tsx
@@ -5,8 +5,8 @@ import {trackAnalytics} from 'sentry/utils/analytics';
 import {generateLinkToEventInTraceView} from 'sentry/utils/discover/urls';
 import useOrganization from 'sentry/utils/useOrganization';
 import {normalizeUrl} from 'sentry/utils/withDomainRequired';
+import {SPAN_ID_DISPLAY_LENGTH} from 'sentry/views/insights/http/settings';
 import type {ModuleName} from 'sentry/views/insights/types';
-import {SPAN_ID_DISPLAY_LENGTH} from 'sentry/views/performance/http/settings';
 
 interface Props {
   location: Location;
