+++ static/app/views/insights/http/charts/responseCodeCountChart.tsx
@@ -2,7 +2,7 @@ import {t} from 'sentry/locale';
 import type {Series} from 'sentry/types/echarts';
 import Chart, {ChartType} from 'sentry/views/insights/common/components/chart';
 import ChartPanel from 'sentry/views/insights/common/components/chartPanel';
-import {CHART_HEIGHT} from 'sentry/views/performance/http/settings';
+import {CHART_HEIGHT} from 'sentry/views/insights/http/settings';
 
 interface Props {
   isLoading: boolean;
