+++ static/app/views/insights/http/charts/responseRateChart.tsx
@@ -8,7 +8,7 @@ import {
 import Chart, {ChartType} from 'sentry/views/insights/common/components/chart';
 import ChartPanel from 'sentry/views/insights/common/components/chartPanel';
 import {DataTitles} from 'sentry/views/insights/common/views/spans/types';
-import {CHART_HEIGHT} from 'sentry/views/performance/http/settings';
+import {CHART_HEIGHT} from 'sentry/views/insights/http/settings';
 
 interface Props {
   isLoading: boolean;
