+++ static/app/views/insights/http/httpLandingPage.spec.tsx
@@ -6,7 +6,7 @@ import {render, screen, waitForElementToBeRemoved} from 'sentry-test/reactTestin
 import {useLocation} from 'sentry/utils/useLocation';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
-import {HTTPLandingPage} from 'sentry/views/performance/http/httpLandingPage';
+import {HTTPLandingPage} from 'sentry/views/insights/http/httpLandingPage';
 
 jest.mock('sentry/utils/useLocation');
 jest.mock('sentry/utils/usePageFilters');
