+++ static/app/views/insights/http/queryParameterDecoders/panel.tsx
