+++ static/app/views/insights/http/queryParameterDecoders/responseCodeClass.tsx
