+++ static/app/views/insights/http/referrers.tsx
