+++ static/app/views/insights/http/settings.ts
