+++ static/app/views/insights/http/tables/domainCell.tsx
@@ -5,7 +5,7 @@ import Link from 'sentry/components/links/link';
 import {space} from 'sentry/styles/space';
 import {useLocation} from 'sentry/utils/useLocation';
 import {OverflowEllipsisTextContainer} from 'sentry/views/insights/common/components/textAlign';
-import {NULL_DOMAIN_DESCRIPTION} from 'sentry/views/performance/http/settings';
+import {NULL_DOMAIN_DESCRIPTION} from 'sentry/views/insights/http/settings';
 import {useModuleURL} from 'sentry/views/performance/utils/useModuleURL';
 
 interface Props {
