+++ static/app/views/insights/http/tables/domainsTable.tsx
@@ -18,8 +18,8 @@ import useOrganization from 'sentry/utils/useOrganization';
 import {renderHeadCell} from 'sentry/views/insights/common/components/tableCells/renderHeadCell';
 import {QueryParameterNames} from 'sentry/views/insights/common/views/queryParameters';
 import {DataTitles} from 'sentry/views/insights/common/views/spans/types';
+import {DomainCell} from 'sentry/views/insights/http/tables/domainCell';
 import {ModuleName, type SpanMetricsResponse} from 'sentry/views/insights/types';
-import {DomainCell} from 'sentry/views/performance/http/tables/domainCell';
 
 type Row = Pick<
   SpanMetricsResponse,
