+++ static/app/views/performance/landing/widgets/components/selectableList.tsx
@@ -14,11 +14,11 @@ import {getConfigurePerformanceDocsLink} from 'sentry/utils/docs';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
 import {NoDataMessage} from 'sentry/views/insights/database/components/noDataMessage';
-import {CACHE_BASE_URL} from 'sentry/views/performance/cache/settings';
 import {
   MODULE_DOC_LINK,
   MODULE_TITLE as HTTP_MODULE_TITLE,
-} from 'sentry/views/performance/http/settings';
+} from 'sentry/views/insights/http/settings';
+import {CACHE_BASE_URL} from 'sentry/views/performance/cache/settings';
 import {getIsMultiProject} from 'sentry/views/performance/utils';
 
 type Props = {
