+++ static/app/views/performance/utils/useModuleURL.tsx
@@ -1,12 +1,12 @@
 import useOrganization from 'sentry/utils/useOrganization';
 import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import {BASE_URL as DB_BASE_URL} from 'sentry/views/insights/database/settings';
+import {BASE_URL as HTTP_BASE_URL} from 'sentry/views/insights/http/settings';
 import {ModuleName} from 'sentry/views/insights/types';
 import {BASE_URL as AI_BASE_URL} from 'sentry/views/llmMonitoring/settings';
 import {BASE_URL as RESOURCES_BASE_URL} from 'sentry/views/performance/browser/resources/settings';
 import {BASE_URL as VITALS_BASE_URL} from 'sentry/views/performance/browser/webVitals/settings';
 import {BASE_URL as CACHE_BASE_URL} from 'sentry/views/performance/cache/settings';
-import {BASE_URL as HTTP_BASE_URL} from 'sentry/views/performance/http/settings';
 import {BASE_URL as APP_STARTS_BASE_URL} from 'sentry/views/performance/mobile/appStarts/settings';
 import {BASE_URL as SCREEN_LOADS_BASE_URL} from 'sentry/views/performance/mobile/screenload/settings';
 import {BASE_URL as MOBILE_UI_BASE_URL} from 'sentry/views/performance/mobile/ui/settings';
